services/boardroom-sync/requirements.txt
watchdog>=3.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
